Simple yet powerful new role for Treex nodes, bundles and documents, which allows to store 'wild' attributes with these objects -- attributes that are not declared in the PML schema. A hashref for each of these objects is returned by its wild() method. Anything what is stored into this hash is serialized into the wild_dump attribute (plain string) by Data::Dumper when storing the document, and reconstructed by eval when loading the document.
All you have to do if you want to store such a new wild attribute is to fill it:

$node->wild->{name_of_my_tentative_attribute_which_I_need_to_store} = $value;

Expected use cases: you need to store some data structures which are not defined 
by the Treex PML schema because (1) you do not want to change the schema
(e.g. the new attributes are still very unstable, or they are likely to serve only
for tentative purposes, or you do not feel competent to touch the PML schema), or (2)
you cannot change the schema, because you do not have write permissions for the
location in which Treex::Core is installed.

This role might be later added to DocZone and BundleZone if needed.



git-svn-id: https://svn.ms.mff.cuni.cz/svn/tectomt_devel/trunk/treex@6193 7cf035df-8472-40d0-8fbe-a733a24d2855