By pressing 'w' in TrEd you can wrap (or unwrap) certain types of trees.
By wrapping I mean that only tree root will be displayed instead of the whole tree.

After skipping to other bundle, trees of the same level located in a zone
with the same language and selector will be wrapped (or unwrapped) too.

This shortcut should be useful especially when bundles contain more than a few
trees, and using grid placement (shortcut 'c') does not lead to satisfactory look. 


git-svn-id: https://svn.ms.mff.cuni.cz/svn/tectomt_devel/trunk/treex@8435 7cf035df-8472-40d0-8fbe-a733a24d2855