New feature for wildcard expansion of filenames - usable in all readers derived from BaseReader.   treex -Len Read::Treex from='!dir??/file*.treex' my.scen   treex -Len my.scen -- '!dir??/file*.treex' # same as above   treex -Len my.scen -- dir??/file*.treex    # also same, but here the expansion is done by Bash, not by Perl, so you are limited by max command length
Moreover,  you can combine it with other ways
  treex -Len my.scen -- '!*.treex @my.filelist another_file.streex'

Therefore, we don't need "treex --glob" anymore. It wasn't used much (except for 4 tests) anyway.
It was not applicable to all readers. It complicated the Run.pm code.

BTW: The @filelist and !wildcard conventions are used in several other tools, e.g. 7z or javac.


git-svn-id: https://svn.ms.mff.cuni.cz/svn/tectomt_devel/trunk/treex@9025 7cf035df-8472-40d0-8fbe-a733a24d2855