Appositions in HamleDT (so far only for cs and grc) are treated as hypotactic structures marked with a new afun.
In PDT, apposition is treated as a paratactic structure (similarly to coordinations) governed by a comma with afun=Apos.
In HamleDT, the first member of apposition governs the (optional) comma and the second member which has afun=Apposition.
We cannot reuse afun=Apos for HamleDT because it would break several tests and assumptions.


git-svn-id: https://svn.ms.mff.cuni.cz/svn/tectomt_devel/trunk/treex@9444 7cf035df-8472-40d0-8fbe-a733a24d2855