Now you can use $node->remove({children=>'rehang_warn'});
It is quite often that we need to delete a node, but rehang its children to parent, in order to not loose them.
It is also common that we such situation is rare/suspicious and we want to produce a warning that the node had children (while we expected no children).
The default behavior remains the same as before: children=>'remove' (without any warning).


git-svn-id: https://svn.ms.mff.cuni.cz/svn/tectomt_devel/trunk/treex@14270 7cf035df-8472-40d0-8fbe-a733a24d2855