Block parameter auxk_to_root=0 results in not calling attach_final_punctuation_to_root().
I thought this may be useful for Spanish, which uses opening question mark ¿ and exclamation mark ¡,
so it can easily mark that only a part of sentence (clause, direct speech) is enclosed in the marks.
By attaching AuxK to root we make the paired marks to have different parent.

Unfortunatelly, the Spanish parser does not always get the marks to the same parent, so I am not going to use this parameter now.


git-svn-id: https://svn.ms.mff.cuni.cz/svn/tectomt_devel/trunk/treex@14275 7cf035df-8472-40d0-8fbe-a733a24d2855