Ondrej Dusek suggested to detect also URLs not preceded by a space.
URI::Find::Schemeless cannot detect URLs which are not preceded by a space (or beggining of string).
Various types of brackets were already separated, but it is possible to find e.g. "go to:http://mysite.com"
without the space after colon (or comma, or other word). Let's detect such URLs as well.


git-svn-id: https://svn.ms.mff.cuni.cz/svn/tectomt_devel/trunk/treex@14511 7cf035df-8472-40d0-8fbe-a733a24d2855