no need to use Write::XY substitute={bar}{foo} to=. now you can use just Write::XY substitute={bar}{foo}
The output filename for text writers is "-" (STDOUT) by default.
to=. means that you want to override this default to the name of the document (based on the name of the original input file).
However, when substitute option is given, you always want to save it into a file (not to STDOUT), even for text writers.


git-svn-id: https://svn.ms.mff.cuni.cz/svn/tectomt_devel/trunk/treex@14652 7cf035df-8472-40d0-8fbe-a733a24d2855