better handling of invalid scenarios
Let's say Scen::XY::get_scenario_string
returned a string which cannot be parsed as a scenario,
e.g. "W2A::Foo param=forgotten quotes"
Before this commit, this resulted in
$ treex --dump A2A::Bar Scen::XY A2A::Baz
 A2A::Bar
 Scen::XY
 A2A::Baz
which means that Scen::XY was not expanded
and when trying to execute it treex failed
because it was looking for a *block* Treex::Block::Scen::XY.
This is a strange result of $return=undef in Parse::RecDescent.
Now, we have an informative error message.