remember also attributes not set by $args
For example, Treex::Scen::EN2CS has sub BUILD, which does
     if (!defined $self->gazetteer){
       $self->{gazetteer} = $self->domain eq 'IT' ? 1 : 0;
     }
So even when "gazetteer" was not passed to new({domain=>'IT'}),
it is now available as $self->gazetteer.
We want to propagate this parameter to subscenarios,
so we want it to be included in $self->args_str.
I consider this solution the least non-elegant one:-).