Fixing DocumentReader to work with skip_finished in parallel
In parallel processing, using skip_finished could lead to unpredictable
results as each of the workers creates its own list of files to
process. The lists will not be identical if some of the workers start
earlier and finish some files before other workers start. This leads
to some files being processed multiple times (with clashes on the output,
making it unreadable), and some files not being processed at all.

This is an attempt to fix this by overriding the file lists in parallel
processing, replacing the list with just the single file to be processed
next each time.