Simple coreference evaluation
Coreference evaluation similar to the one proposed by Tuggener (2014).
In fact, this type of evaluation has been used at UFAL since very beginning,
called by no or misleading names (pairwise F-measure).

For each anaphor candidate, its (non-)linkage in the gold annotation
is compared with the system annotation. This results in four different
categories: true positives (gold and system linkage agree), false
positives (no gold linkage, some system linkage), false negatives
(no system linkage, some gold linkage), and wrong linkage (gold and
system linkage disagree).

Unlike in (Tuggener, 2014) a gold and system linkage agree if the system
antecedent belong to the entity defined by the anaphor. In other words,
the antecedent may be any of the mentions belonging to the entity.

This block also has to deal with possible discrepancies between the
gold and system annotation not related directly to the coreference.
For instance, automatic tokenization differing from the gold tokanization
may prevent finding the system counterpart of a gold mention. The same
happens with differences in what is represented on the t-layer as
an individual node as well as with the nodes newly generated on the
t-layer.

The output follows a three-column format with the individual columns
representing the following:
1) is there a gold link?
2) is there a system link?
3) do the links agree? (they agree also when there is neither gold nor a system link)

This output file can be turned into accuracy, precision, recall,
and f-score measures by running the scripts/eval.pl from the MLyn repository.

TODO: add documentation => this text + usage