Splitting Scen::Analysis::CS into separate scens for M, N, A and T.
This is useful because not everyone needs the full t-layer analysis,
so I suggest splitting the large scenarios in this way.
I am now only applying this to Analysis::CS,
but if we agree that this is something we want to do,
we can apply it to other scens as well.