fixed a bug where overriding process_start prevented process_end to run
The problem was that
 after 'process_start' => sub {};
in Treex::Core::Block
is ignored if a block overrides process_start.
So instead of using the "after" hook,
let's set $block->is_started each time we start the block.
I am not sure why the is_started is need,
but perhaps there is a reason.