Treat "ne" and "nikoli(v)" RHEM/CM similarly to #Neg
As TFA is not annotated outside of PDT, the treatment was simplified:
Neg.RHEM always negates the parent, ne.RHEM always negates the right
brother, and CM always negates the elements of its coap parent.

Also, add "ne" and "nikoliv" to the alignment of the negated word
after removing them.